feat(UA): Add columns to DBML/SQL
UserApis: Added recipients column image_url to store the associated image URL (if any).  Added tags column category to store the category for the tag (if any).  Added recipient_tags columns weight to store the relevance weight for this tag with the recipient and added_at to track when the tag was last added.