feat(AP): Add recipient/message/tag properties
AddProperties: Updated model classes with new properties.  Added Recipient.tagsLastGenerated to hold the date/time the recipient's tags were last generated based on form data.  Added Message.donorRead to indicate if the message has been read or not.  Added RecipientTag.selected column to indicate if the recipient has selected that tag or not.